roundtable_dockerize/keyword_api/register_login/requirements.txt
click
h11
pydantic
starlette
fastapi
PyJWT
passlib
uvicorn
python-dotenv
pandas